net: asix: Fix ASIX 88772B with driver model
Commit 0d4d49f42f1a ("net: asix: fix operation without eeprom")
added a special handling for ASIX 88772B that enable another
type of header. This break the driver in DM mode as the extra handling
needed in the receive path is missing.

However this new header mode is not required and only seems to
increase the code complexity, so this patch revert this part of
commit 0d4d49f42f1a.

This also reverts commit 21be5776ebfc65cf37826bb3291a5bbb3f991065
("net: asix: Fix AX88772B when used with DriverModel") of late.

Fixes: 0d4d49f42f1a ("net: asix: fix operation without eeprom")

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>